rabbit_ff_registry: Refactor as a `persistent_term`
Previously the feature flag registry was a module which was regenerated
(i.e. with `erl_syntax` and `compile:forms/2`) whenever the set of known
feature flags or feature flag states changed. Compiling and loading
a module is fairly expensive though, costing around 35-40ms for me
locally, and the registry module is regenerated many times on boot. So
this change saves a fair amount of time on boot. A regular single-node
broker boots locally for me in around 4250ms via `bazel run broker`.
With this patch the boot time falls to around 3300ms.

This patch also improves the time it takes to enable individual feature
flags. `time rabbitmqctl enable_feature_flag khepri_db` for example
changes locally for me from around 1.2s to 1.0s with this change.