Replicate x_jms_topic_table Mnesia table
The x_jms_topic_table Mnesia table must be on all nodes
for messages to be published to JMS topic exchanges
and routed to topic subscribers.

The table used to be only in RAM on one node, so it would
be unavailable when the node was down and empty
when it came back up, losing the state for subscribers
still online because connected to other nodes.

References #9005

(cherry picked from commit df9fec88e360a679cf489b57a1187c13dd342f42)

# Conflicts:
#	deps/rabbitmq_jms_topic_exchange/src/rabbit_db_jms_exchange.erl