Ra 2.10.0
This Ra release contains a number of fixes and improvements including:

* Much improved resiliency when Ra infrastructure such as the WAL or
segment writer encounters unexpected errors during disk operations.

It also includes the following features that are RabbitMQ does not
yet make use of (but will in the near future).

* Checkpoints: allow non truncating snapshots to be written
to allow faster recovery of quorum queues with long backlogs for example.
* Server recovery strategy configuration: allow dynamically started
ra servers to be optionally restarted.
* New handle_aux/5 callback with a better and safer API