rabbit_ldap_seed: Remove 'memberOf' annotation for "peter" seed user
In openldap 2.5, `memberOf` seems to be a read-only annotation. Instead
the LDAPWiki recommends adding the user's name to the member attribute
on the relevant group objects:
<https://ldapwiki.com/wiki/Wiki.jsp?page=MemberOf>

We already do that when setting up the group objects, so this annotation
is safe to remove.

(cherry picked from commit 8c5cf257df56ed3dbcaf50a8ebc7a55c5701f458)