rabbitmq_peer_discovery_consul: Separate service name and ID
[Why]
The `consul_svc` parameter is used as the service name and to construct
the service ID. The problem with the way the service ID is constructed
is that it doesn't allow to register several distinct RabbitMQ nodes in
the same Consul agent.

This is a problem for testsuites where we want to run several RabbitMQ
nodes on the same host with a single local Consul agent.

[How]
The service ID has now its own parameters, `consul_svc_id`. If this one
is unset, it falls back to the previous construction from the service
name. This allows to remain 100% compatible with previous versions.

(cherry picked from commit 684ec76f77083f87ab32b230fa31283a99584957)