rabbitmq_peer_discovery_consul: Populate Erlang node name during registration
[Why]
This allows other nodes to discover the actual node names, instead of
deriving one from the Consul agent node name and their own node name.

This permits to register several RabbitMQ nodes in the same Consul
agent. This is at least handy for testsuites.

[How]
The Erlang node name is added to the `Meta` properties list as long as
the RabbitMQ cluster name.

Note that this also fixes when the cluster name is added to `Meta`:
before this commit, a non-default cluster name was not added if the
user-configured properties list was empty at the beginning.