Don't show disabled plugins as pending upgrade
Since commit c0187ec15 the value of `running_version` is converted
to_string (`nil` would become the empty string). But the formatter
expected `running_version` to be `nil` if the plugin is not
running. This did not match and not running/not enabled plugins were
marked incorrectly as "pending upgrade to...".

For example:
```
$ rabbitmq-plugins list trust -q
[  ] rabbitmq_trust_store               (pending upgrade to 3.13.0+51.g9145b53)

$ rabbitmq-plugins list trust --formatter erlang -q
 #{status => running,format => normal,
  plugins =>
      [#{enabled => not_enabled,name => rabbitmq_trust_store,running => false,
         version => <<"3.13.0+51.g9145b53">>,running_version => <<>>}]}
```

(cherry picked from commit 69eb2b9c299587f71806caaa26b95fc9361d587c)