make: Don't immediately expand PROJECT_VERSION
When there is nothing to do we don't need this variable
so we don't want to calculate it unnecessarily.

Because this variable is only used once, when
producing the .app file, we don't have to worry
about the calculation being done multiple times.
If we ever do then it will need to be lazily
evaluated[1] instead.

[1] Managing Projects with GNU Make, 3rd Edition Chapter 10

Execution speed differences:

  make -C deps/rabbit nope  0,02s user 0,03s system 101% cpu 0,051 total
  make -C deps/rabbit nope  0,02s user 0,01s system 97% cpu 0,031 total