make: Easy code reloading for run-broker/start-cluster
This enables code reloading for single nodes and
for clusters. For single nodes using `make run-broker`
at least two terminals must be used: the one with
the broker, and the one doing the reloading.

When only using a single node (`make run-broker`),
the `RELOAD=1` variable can be passed to Make when
recompiling. This will lead to code reloading at
the end of compilation.

  make RELOAD=1

Alternatively the "reload-broker" target can be used:

  make all reload-broker

Special care must be given when the -j flag is used.
In that case it may be needed to separate compilation
and reloading in two separate calls:

  make && make reload-broker

The same considerations apply to clusters, only there
isn't a shorthand variable for them. Instead, the
"reload-cluster" target must be used:

  make all reload-cluster

Or:

  make && make reload-cluster

Either of these will compile and reload the modules
on all nodes of the cluster.

It is also possible to reload modules only on some nodes,
using `make reload-broker` with the variable "RABBITMQ_NODENAME"
set to the node you want to trigger a reload on. This could
be used to test scenarios where the code differs between
the nodes in the cluster.