Remove +warnings_as_errors from the jose compilation options
Why?
1. Generally +warnings_as_errors is unnecessary for deps anyway
2. The gazelle plugin in rules_erlang, for the sake of simplicity,
splits compilation of a given application into at most 3 phases,
parse_transforms, behaviours & everything else. Therefore, if there
are behaviours that are behaviours, this produces warnings. It's
certainly possible to write an extra phase into BUILD.jose so that the
warning does not occur, but in this case it's much simpler to just to
allow the warning. Futhermore, this should not be an issue with
rules_erlang 4, as it is not limited to the finite number of
compilation phases, nor does it even have a gazelle plugin.

(cherry picked from commit fbf896c05cfbe2742ec93120f6b2f50319a23de8)