rabbit: Stop `rabbitmq_prelaunch` if we fail to start `rabbit`
[Why]
This is important if the environment changes between that error and the
next attempt to start `rabbit': the environment is only read during the
start of `rabbitmq_prelaunch' (and the cached context is cleaned on
stop).

This fixes a transient failure of the
`unit_log_management_SUITE:log_file_fails_to_initialise_during_startup/1`
testcase.

V2: Explicitly ignore the return value of `application:stop/1`.