make: Don't start all when run-broker a specific plugin
When FAST_RUN_BROKER=1 was introduced it helped reduce the
time to run the broker, but mistakenly was always starting
all plugins even when running run-broker against a specific
plugin, for example `make -C deps/rabbitmq_management run-broker`.

  Starting broker... completed with 36 plugins.

With this commit only the target plugin (and its dependencies)
will be started.

  Starting broker... completed with 3 plugins.

This also has a positive effect on start performance:

  make -C deps/rabbitmq_management run-broker  2,28s user 2,11s system 88% cpu 4,943 total
  make -C deps/rabbitmq_management run-broker  2,00s user 1,61s system 94% cpu 3,807 total