Ra v2.11.0
This release includes a new machine API `snapshot_installed/2`. This new
API will only be used indirectly through khepri.

This release also includes an performance improvement that reduces the chances
of building a large WAL mailbox backlog when a node is low on scheduling
resources and commands are committed by followers completing writes to disk
before the leader.

There is also a fix for a potential election deadlock.